Slight change to session clean up to leave sessions inactive for at least 24 hours.

+++ jadaptive-boot/src/main/java/com/jadaptive/app/session/SessionCleanupTask.java
@@ -1,5 +1,8 @@
 package com.jadaptive.app.session;
 
+import java.util.Date;
+
+import org.apache.commons.lang.time.DateUtils;
 import org.pf4j.Extension;
 import org.springframework.beans.factory.annotation.Autowired;
 
@@ -7,6 +10,7 @@ import com.jadaptive.api.permissions.PermissionService;
 import com.jadaptive.api.scheduler.ScheduledTask;
 import com.jadaptive.api.session.Session;
 import com.jadaptive.api.session.SessionService;
+import com.jadaptive.utils.Utils;
 
 @Extension
 public class SessionCleanupTask implements ScheduledTask {
@@ -24,7 +28,11 @@ public class SessionCleanupTask implements ScheduledTask {
 		
 		try {
 			for(Session session : sessionService.inactiveSessions()) {
-				sessionService.deleteSession(session);
+				
+				Date threshold = DateUtils.addDays(session.getSignedOut(), 1);
+				if(threshold.before(Utils.now())) {
+					sessionService.deleteSession(session);
+				}
 			}
 		} finally {
 			permissionService.clearUserContext();
